Add files via upload
Updating IndexInsta.html: I have embedded three instagram posts which link to the wildlife related profiles. Struggled to embed a Instagram Feed so I have done a workaround of print screening an image of a feed and href linking the image to the #Londonwildlife. There are also two images that take you to other wildlife related Instagram accounts.