Add files via upload
Updating indexgames.html: Changes include Integrating foxy into the page. I have embeded a screen shot image of the game and href linked it to download the game as a zip file. Need help with the formatting of the page as foxy run is right at the bottom with a big gap on the page. This could be temporary until we find a better way to display this game!